quota: make handling of each global quota fully atomic
We now first check all the local quotas, then atomically check and charge
each global quota in turn, before charging the local quotas.

We still risk having to revert one global quota if a different one fails,
but now this should be much less likely, and we never charge a global quota
if it is going to fail.

Signed-off-by: Tom Gundersen <teg@jklm.no>