user: convert peer_info counters to count remaining resources
Convert the peer_info resource counters to work like the user resource
counters: Count the number of remaining resources, rather than the number
of used resources.

This has the advantage that we only need to touch a single counter at
runtime. There is no reason to even read the maximums. Furthermore, we can
drop the maximums from peer_info and move them into the user object.

In the future, the logic is as follows: A user accounts for all global
resources of any peer of that UID. It has default maximums, which are
used to initialize the inverse-counters. However, we also remember the
maximums, so if we add an ioctl that changes the maximums of a user, we
can calculate the delta and modify a user object.

Whenever you create a new peer, we initialize the maximums of that peer
based on the maximum of the user, except if you provide your own maximums
via the INIT/CLONE ioctls (which we currently do for pool-size *and*
implicitly via RLIMIT_NOFILE).

Note that this patch disconnects the pool.size from peer_info.n_allocated.
This might have to be disconnected further later on, by making the pool
dynamically sized, ignoring any framing overhead, and solely rely on
peer_info.n_allocated.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>