transaction: fix handling of erroneous messages
Whenever a message has no slice set, its sole purpose is to serve as
n_dropped handler in a transaction. This means, the message is still
queued, ordered, and properly used, but should be dropped at commit time
and turned into an atomic_inc(&destination->n_dropped).

Fix the handling in bus1_transaction_consume() to properly handle such
messages the same way as normal ones, with the exception that they will
never be committed in the queue. So far, we didn't properly unlink them
from the queue, nor sync/tick the clocks.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>