handle: fix bus1_handle_inflight_install()
Don't use a sophisticated (and broken) logic to group handle-installs.
Skip that for now, just install all handles as 3rd party handles an be
done. This works, and if we need performance enancements for batch handle
installs, do that later on.

This just drops the n_new_local logic and installs everything as 3rd party
handle. The n_new_logic is completely broken.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>