handle: make queue-link own its own reference
We used to attribute the queue-link reference to the "committed" state of
the queue-node. That is, only if the node is committed, it owns a
reference. However, this breaks if a peer is disconnected racing a flush,
because we cannot perform the final notification commit, and as such never
commit the staged notification. As a result, we might end up destroying
the now unused handle without having committed _or_ dequeued it.

We fix this by always attributing its own reference to a _linked_
queue-node. This means, if the notification is staged, but cannot be
committed due to a racing disconnect, then the handle will stay referenced
and linked, and we will no longer pre-emptively destroy it. Instead, we
rely on the peer cleanup to flush staging notifications from the queue
afterwards. This will unlink the node and release its final reference.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>