message: fix n_fds underflow in message install
A loop like this:
  while (n_fds-- > 0)

works fine as long as it is the last time we use @n_fds, since we will
leave it as -1 afterwards. This is fine in the error-path of
bus1_message_install(), but not before. Fix the code to never trigger this
underflow and instead properly split the check and decrement by the
condition.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>