peer: clone - swap to passing handle from parent to child
Currently clone would unconditionally create a new node in the child
and pass a handle to it to the parent. Switch it around so that the
parent instead passes one of its handles to the child.

Currently, we only allow a handle to a on-the-fly-created node to be
passed in, but we will open this up to allowing any handle to be
passed along.

Signed-off-by: Tom Gundersen <teg@jklm.no>