bus1-client: add bus1_client_replace()
This replaces one peer with another. The old peer is freed and the
underlying fd number of the new peer is changed to what the
underlying fd number of the old peer used to be.

This will be useful for passing peer fds over exec, reusing a fixed
fd number.

Note that CLOEXEC is still set for the new peer, so that would
have to be cleared before actually passing the fd over fork+exec.

Signed-off-by: Tom Gundersen <teg@jklm.no>