handle: add uapi for owner-release notification
This is prep-work for owner-release notifications, which we intend to send
out when all handles for a node have been released. Right now we
immediately destroy the node, but this breaks compatibility with existing
IPC. Instead, we want nodes to always stay around unless explicitly being
destroyed. But this requires us to notify node owners of nodes becoming
unused. This commit adds the uapi for this, but does not hook it up, yet.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>