peer: lift allocation restriction on BUS1_CMD_PEER_CLONE
So far we required users to pass BUS1_NODE_FLAG_ALLOCATE via
BUS1_CMD_PEER_CLONE. This was for historic reasons. This patch lifts that
restriction and allows any input handle to be passed.

The @parent_handle field of BUS1_CMD_PEER_CLONE now works exactly like
auxiliary handles in message transactions. That is, it is normally used as
input to specify the handle to use, but if BUS1_NODE_FLAG_ALLOCATE is
given, it will allocate a new node, return its ID to the caller and also
acquire a new user-ref (this user-ref is *only* acquired if a node was
allocated).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>