peer: cache waitq pointer in peer-info
We need the waitq inside of the real peer structure as poll() callbacks
cannot easily deal with vanishing wait-queues. However, the waitq is
required all over the place but we not necessarily know about the parent
peer object in all contexts. This makes dealing with wake-ups quite
annoying.

Cache a back-pointer to the wait-queue inside of the peer-info so we can
easily wake up a peer at any time without requiring inside knowledge of
teardown paths were acquisition of active refs might fail.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>