queue: introduce separate queue-lock
Right now we have to jump through hoops to do synchronous handle
destructions. We must stage and commit all destruction notifications on
any handle owner in a synchronous manner. This requires us to unlock the
owner, queue everything (possibly 'soft-live-locking' the handle), lock
the owner again and commit. This requires us to be aware that *any*
handle-release might re-lock a peer.

This looked easy in the beginning but caused the code to be almost
unreadable. This commit introduces a separate queue lock on each peer,
which just protects the queue. This allows us to keep a peer locked while
modifying a queue.

We already have the queue implementation completely independent of the
parent peer. Hence, this is quite trivial to introduce. Follow-up patches
will carefully simplify the handle logic to reduce the amount of state
tracking that is needed right now. Eventually, this will allow us to turn
bus1_handle_stage_relock() into a synchronous bus1_handle_stage() and
thus remove the 'relock' notion we have all over the place right now
(including removal of 'notify').

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>