handle: extract node-notification
Provide 2 new helpers to queue and dequeue node-notifications. Also hook
them up in the respective places.

For now we hide the queue-operation behind an "if (0)" since we don't want
to commit to the new behavior right now. This will be fixed in a follow up
when tests are in place.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>