peer: print unique peer identification via fdinfo
Since we no longer use our own FS we cannot distinguish peers via fstat().
In fact, due to the inner workings of char-devs, there is no way to
distinguish two FDs to the same char-dev.

This introduces a new 64bit peer-id that is unique per peer. It can be
queried via /proc/<pid>/fdinfo/<fd> and will be printed similar to the
information provided by epollfd, signalfd, eventfd, and friends.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>