peer: provide debug information for each peer
Add basic debugfs infrastructure to see all peers that are active and
print their active-reference state. We get this almost for free, and at
the same time we get runtime lifetime validation via VFS, since it will
complain loudly if we leak resources.

From now on, it will be very easy to provide ad-hoc attributes for
situation based debugging, since all the infrastructure is in place.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>