queue: keep the count of dropped messages in the queue
Conceptually, dropped messages are part of the queue. They are peeked
and dequeued just like regular messages. Move the count from the
peer_info object to the embedded queue object.

At the same time unify our notion of when a queue is readable to always
take dropped messages into account. Earlier, this was done only in
poll().

Signed-off-by: Tom Gundersen <teg@jklm.no>