queue: only allow syncing on even timestamps
We were converting all timestamps to be even, but there is no reason
to ever pass in odd timestamps to begin with. Add a WARN_ON instead.

There was one place where an odd timestamp was passed in, that was
an oversight and is converted to be even without any functional
change.

Signed-off-by: Tom Gundersen <teg@jklm.no>