peer: flush - unify message and notification handling
When flushing, only clean up the payload of committed queue nodes,
and leave the staging nodes to be cleaned up by whomever is in the
process of committing the node.

This is how messages were handled, make notifications the same.

No functional change, but this will be used in the following patch.

Signed-off-by: Tom Gundersen <teg@jklm.no>