transaction: fail transaction if destination has been destroyed
If CONTINUE is passed as SEND flag, the message is still silently
dropped.

However, without CONTINUE we now fail the transaction with
-EHOSTUNREACH in case any of the destination nodes have been
destroyed.

Note that 'before' refers to the local clocks, and in particular
means that the destruction notification of the destination node
would be delivered before the message in question.

Signed-off-by: Tom Gundersen <teg@jklm.no>