peer: send - never ignore ENXIO
Sending a message to an invalid handle is always a local error, not
a remote one, so we should not ignore ENXIO even if the CONTINUE
flag is passed.

It used to be that handle release could be triggered remotely, and
performed asynchronously, but we no longer allow a hanlde to be
invalidated without the knowledge of the local peer (either through
the receipt of a node destruction notification or by actively
releasing the handle).

Signed-off-by: Tom Gundersen <teg@jklm.no>