handle: split off batch_release from batch_destroy
Releasing a handle might require locking the owning peer (to drop it from
its rb-trees). This is nasty to do in any _destroy(), _free(), or _unref()
function, as it is non-obvious and we generally want to keep those
cleanups as simple as possible.

Split off the handle-release from batch_destroy() so we can clean it up
explicitly on deallocation or explitic release, rather than on final
cleanup.

For now, we just convert the callers to call both, batch_release() and
batch_destroy(), but will convert this in followups.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>