user: quota - fix accounting
We still don't properly account handles.

Any inflight resources are counted globally by the sending user,
locally by the receiving peer and locally by the pair of sending
user/receiving peer.

Once a resources has been dequeued it should no longer be charged
to the sending user, but purely be accounted on the receiving
peer.

Distinguish between bus1_message_deallocate(), which dropps all
quota charges as the message is being dropped, and
bus1_message_dequeue(), which keeps the per-peer charges as the
message is merely dequeued and the resources are still pinned.

Signed-off-by: Tom Gundersen <teg@jklm.no>