handle: bus1_handle_destroy() -> bus1_handle_finish()
The term 'destroy' describes real user-space operations on nodes, so avoid
it for management functions in handle.c. Use _init() as initializer, and
_finish() as deinitializer. This more clearly distinguishes management
from runtime operations.

We would use _start() and _stop() for runtime operations internally, but
that would mismatch with the ioctl names. This is unfortunate, but better
try to be non-ambiguous than 100% uniform.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>