handle: simplify bus1_node_dequeue_notification()
Reduce the weirdness just to avoid bus1_handle_unref() under the
queue-lock. We know that this is not the last unref() so no reason to not
do this under the queue lock. Also, even if it was the last, it is a
pretty cheap operation, given that anything under queue-lock might be
preempty anyway.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>