handle: correctly uninstall owner handle
The rules for owner handles are special. They stay valid in the ID space
even if no inflight ref is taken. An owner must destroy them explicitly to
clear their imaginary keep-alive flag. Once that is done, their lifetime
is the same as normal handles, that is, they're are uninstalled once the
last ref is dropped.

Make sure we do that correctly during owner detach.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>