message: move locking into bus1_message_deallocate()
Like bus1_message_allocate(), keep the locking internal to
bus1_message_deallocate() and avoid requiring callers to take care of
that. This was an optimization in case the caller already holds the lock.
However, this was only applicable for error-paths right now, which we
really don't want to optimize anyway.

We can always provide a separate helper in case it turns out it is needed
someday.

This also silently fixes a de-queue inversion bug where we deallocated
before dequeing a message.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>