queue: make queue-nodes ref-counted
We want PEEK operations to work without holding qlock for the entire
operation. The qlock is meant to be as short-lived as possible since it is
taken by remote peers for message transactions. Furthermore, we want to be
able to move it into the queue-implementation to reduce its exposure.

Introduce queue-node ref-counts to allow PEEK operations (and friends) to
retain access to a message *without* taking over ownership.

This just introduces the ref-count but does not introduce any functional
change. PEEK and friends are left untouched for now.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>