peer: accounting - don't limit the number of slices or handles
This used to be set to the max per-user value, but that has the
same effect as not having any limit as the quota logic takes
the minimum of the per-user and the per-peer limit.

Moreover, if the per-user limits becomes dynamic in the future,
copying it to the peer like this would not work as expected.

No functional change.

Signed-off-by: Tom Gundersen <teg@jklm.no>