ioctl: replace PEER_CLONE with HANDLE_TRANSFER
Rather than implicitly creating a new peer in order to transfer a
handle between two disconnected peers, allow a direct transfer
between two peer fds.

This means that connections between previously disconnected islands
of peers can be made. However, this is not a significant change in
behavior as it is virtually indistinguishable from having a peer
work as a proxy between the two islands.

This further simplifies the creation of peers, to only be performed
by opening the right device node.

Signed-off-by: Tom Gundersen <teg@jklm.no>