queue: fix incorrect front assignment on re-stage
Whenever we re-stage a queue entry, we must check whether any hidden entry
next to it might become the new queue front. That is, if our entry will be
re-staged *after* that next entry, and the next entry is committed, it
will become the new front.

Right now we use the *old* ordering keys of our entry, rather than the new
to-be-set-timestamp to check whether the queue node will be inserted after
it. This is wrong.

Change our compare-function to compare a node to a given key, rather than
another node, and then pass the correct key in all situations. While at
it, expose the core of it to be used in handle.c notification ordering.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>