handle: avoid old userref-publish in handle-inflight
The handle-inflight object is the last user of the old userref-publish
helper, which is completely broken on dead nodes. Convert it to use the
new, safe helper which no longer consumes inflight refs.

This means that the inflight refs of handle-inflight objects remain valid
until the final destruction of the object, which already properly releases
them.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>