peer: fill in msg-info in bus1_peer_queue_peek()
We want to unify the different message types, hence, fill in the message
information directly when fetching it from the queue. For now, we always
fill in the 'data' portion, even though the following handlers still use
the different message types. This will be fixed in a followup. Luckily,
the 'destination' and 'handle' fields of the different message types
overlap, anyway, so tests could already switch to the new type.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>