message: stop using message->data.n_fds
We want to get rid of message->data, since it is not needed in its
entirey. Hence, start moving single fields which *are* required back into
the message structure. This commit starts with data.n_fds and replaces it
with message->n_files. Furthermore, we avoid the 64bit integer and rather
use the more correct size_t.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>