message: avoid using message->data.n_bytes
We want to get rid of the redundant message->data field, so move the
required 'n_bytes' field out into a separate 'size_t', similar to n_files.

Note that 'n_bytes' is *not* equivalent ot slice->size, since the slice
size is aligned *and* additionally may account for FDs and handles.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>