message: drop message->data
The message->data field was always a hack to avoid copying all fields
manually. However, it is a waste of memory, since it is partially
redundant and uses API types, rather than smaller native types.

We converted most of the users of message->data already, this commit moves
the remaining trivial fields into 'message' directly as well, and then
drops the 'data' member.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>