bus1: introduce BUS1_RECV_FLAG_INSTALL_FDS
File-descriptors are bound to a process, therefore, their IDs will differ
if multiple different processes PEEK at the same message on a shared peer.
A caller must be aware of this and provide sufficient synchronization. To
give them the right tools, this adds BUS1_RECV_FLAG_INSTALL_FDS. If the
flag is not given, the FD-numbers in the slice are left untouched by RECV,
nor are any FDs intalled into the receiver's process. But if the flag is
given, the FDs are installed *and* the slice is modified on *every*
invocation of RECV with this flag set.

Note that even a failed RECV operation might modify the slice, since we
have to copy the FD numbers into the slice before installing them. If that
fails mid-way through, we will return an error, but might have modified
the slice already. Hence, if you use that flag, you must be aware that the
slice FD-content is now invalid, until this call succeeds. This should be
fairly obvious, though, since the call is non-atomic anyway, and it will
very likely return different FD information, anyway. Hence, a racing
reader of the FD-content will be buggy, anyway.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>