handle: support transmitting invalid handles
We explicitly support lazy initialization, so user-space might be aware
that a specific handle does not exit, but still wants to perform a SEND,
to make sure it looks the same as if the handle was invalidated in
parallel. This is very important to make sure data always looks the same,
regardless whether you do lazy initialization or not.

Make sure we allow BUS1_HANDLE_INVALID as transfer handler, and transmit
it as if the handle got invalidated before.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>