queue: re-order nodes before queues
This just moves code-around, no functional change whatsoever.

This moves all the node-handling *before* the queue-handling. This annoyed
the hell out of me several times. The node handling does never depend on
the queue, so it should logically be defined *before* the queue handling
is.

This avoids quite a lot of forward-declarations and/or weird function
ordering in queue.c.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>