handle: properly account nodes on a peer
For each node that is created, decrement peer_info->user->n_handles. Since
we implicitly say that a node always pins a handle (as long as it is
live), we must make sure those nodes are accounted on the owning user.

This commit *only* accounts the handles that are pinned by a node as long
as it is live. It does not affect any other handle reference.

To properly account nodes, we must hook into any ioctl that might call
bus1_handle_attach_owner() (i.e., bus1_handle_new_owner()), or
bus1_node_destroy().

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>