user: move n_handles hack into message handling
Rather than making the user code work around missing handle-accounting,
move that logic into message.c. This is the place we were actually do the
resource conversion, so it is also the place that will be fixed once
handle accounting is available.

This makes user.c unaware of any remaining TODO items, and keeps them
contained in message.c instead.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>