docs: update bus1.7
This adds extensive information to bus1.7 and tries to make it more
similar to epoll.7, signal.7, eventfd.2, signalfd.2, timerfd_create.2, and
friends.

We really want this man-page to follow the style of other linux man-pages,
so lets not invent our own naming scheme, but stick to what is common.

This means:
  No self-invented refsect1. Put all real description text into
  DESCRIPTION. Use sub-sections for further structuring.

  Use the same naming-scheme as other similar existing man-pages. Don't
  invent our own tables, or wording. Stick to what is already there!

  Don't invent our own man-page names. There are 7 kinds of man-pages:
  overview pages, and syscall pages. Lets not invent bus1.foobar.7, they
  do not fit the bill. Instead, write a proper bus1.7 overview, and
  everyone will be happy.
  Yes, the page will get very long, but we will see 2 benefits: The page
  is better maintained, since you don't have to look at 10 different pages
  on each change you do. Secondly, we will keep the bus1 API short, so the
  bus1.7 overview is actually *able* to give a thorough overview.
  (I might be dreaming..)

This adds some structure and initial text to bus1.7. The old man-pages are
left for now, but will be removed once properly converted. Note that I
would actually really prefer bus1_client_foobar(3) pages as well. Those
can be more detailed regarding user-space usage, examples, etc. But for
now, lets keep kernel semantics in bus1.7 and keep them technical, but
short.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>