bus1: bump BUS1_VEC_MAX to UIO_MAXIOV
We used to reduce BUS1_VEC_MAX to below UIO_MAXIOV so we can add our own
internal iovecs to the mix. However, this was based on the assumption that
kernel-internal code cares for UIO_MAXIOV. This is not true (only the
importers care), so the argument was flawed right from the beginning.

Furthermore, we no longer add our own iovecs, anyway, so no reason to
divert from UIO_MAXIOV at all. Bump it to stay consistent.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>