handle: fix owner-handle lookup in ioctls
Owner-handles might have n_user==-1, but are still valid in case the node
is live. Introduce a helper that can be used *after* a handle was looked
up by ID (which guarantees the handle was linked at *some* point). It then
checks whether the ioctl is allowed to use it. Two cases exist:

 * The handle is an owner handle, in which case the fact that it was
   linked with its ID is enough to allow its use.

 * The handle is no owner handle, in which case the combination of it
   being linked by ID *plus* the fact that n_user is, at the point of
   checking, elevated, is enough for the handle to be allowed use.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>