message: implement resource pinning
Rather than relying on magic semantics on when, and by whom, message
resources are to be freed, turn this into a pin-counter. Anyone can now
pin the resources, making sure they're available as long as they need it.

Note that this does not make use of it, yet. It simply replaces the old
logic by the new pins.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>