peer: reset - support explicit disconnect
The code for this was always here, but the functionality was not exposed.
With this change a peer can be disconnected even when there are open
file descriptors to it. Any in-flight operation is finished, but any
new operation will fail with -ESHUTDOWN, HUP is signalled by
poll() and all resources are released.

Signed-off-by: Tom Gundersen <teg@jklm.no>