tree-wide: introduce BUS1_WARN_ON
On debug builds this is the same as WARN_ON, but on production
builds this is a noop. This means that it can only be used on
statements without side-effect, but also that we can be more
liberal in the use of BUS1_WARN_ON than we could with regular
WARN_ON.

WARN_ONs with side effects are left untouched, but all others
are converted over.

Signed-off-by: Tom Gundersen <teg@jklm.no>