peer: ioctl - extend node destruction api
Allow a flags argument as well as passing arbitrary number of
nodes. This only changes the api, it does not actually hook it
up.

The flags are needed to optionally allow a node to be destroyed
immediately. I.e., drop all its pending messages and do not queue
a destruction notification for ourselves.

Passing multiple nodes is needed to allow atomic destruction of
nodes that conceptually belong together.

Signed-off-by: Tom Gundersen <teg@jklm.no>