tests: hive - test parallelism
Add the starts of a test to verify correctness of parallel message
transactions.

For now simply create many threads, with one peer each, containing
a node and one parent peer containing handles to each of the children.

Verify that we receive the correct number of node destruction
notifications when destroying each child peer.

Signed-off-by: Tom Gundersen <teg@jklm.no>