handle: rename had_userref to has_userref
The old name made it look like we were checking for whether the
handle ever had a userref, but in fact we are checking if it
currently has one.

The reason for the naming was that we do not guard against the
userref being dropped before using the handle. However, if that
happened it means the HANDLE_RELEASE ioctl must be called
concurrently with whatever does the check, in which case we
can consider the handle to be used before the userref was dropped
witohut anyone being able to tell.

Signed-off-by: Tom Gundersen <teg@jklm.no>