security: implement LSM stubs
Lets move forward with LSM integration and provide stubs for the LSM
hooks. So far they're designed similar to the Binder hooks, but still use
native types and data. They are not set in stone, but are meant to start
the discussion and allow a code-based development of the hooks.

The "security.h" file will be removed once those hooks are provided by
linux/security.h. Until then, we just carry the stubs.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>