transaction: support empty secctx
We have no control over secctx marshaling, so we should not rely on any of
its content. Make sure we transmit the exact same information as UDS,
including a boolean that tells whether the secctx is valid or not
(regardless of its length).

This extends the message-object with a flags field that we return to
user-space. We will use this for further flags in follow-ups.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>