peer: recv - do not dequeue messages that may not be possible to read
Userspace may not have the whole pool mapped at all times, and may
need to remap it if they receive a message outside their mapped
region. However, remapping may fail, and in that case it will not
be possible to correctly clean up the handles and fds from a message
as they cannot be read from the pool.

To avoid this problem, let userspace indicate to the kernel how much
of the pool they are guaranteed to be able to read, and refuse to
dequeue if a message is outside this range. In case RECV fails,
userspace should peek the message to find the necessary pool size,
remap and then RECV again.

Signed-off-by: Tom Gundersen <teg@jklm.no>