queue: make queue_flush() skip staging entries
Rather than returning staging entries to the caller, make sure
bus1_queue_flush() drops them right away. We know that staging entries
always have an owning transaction, so we can rely on them to cleanup the
flushed entries.

Right now, this is not true for notifications. This will be fixed in
follow ups.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>