handle: use single-linked list during destruction
During destruction, all handles are pinned by the ongoing destruction.
This allows iterating them to implement a fast-path transaction. We
always used the double-linked list of the node-link so far, since it was
convenient. However, a single-linked list is sufficient and allows us to
get more temporary storage on the handle structure.

Turn the double-linked list into a single-linked list (just like for
bus1_message handling), and prepare for followups to store more
information in the newly gotten 4-8 bytes.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>