pool: drop lockdep annotations
There are many objects in the kernel that access must be properly
serialized by the caller. It is very uncommon to make the object check
the parent-lock via lockdep. Hence, we should not do this either. Drop
the lockdep annotations in bus1_pool and just tell the caller to lock
properly.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>