transaction: make error handling consistent
This makes error handling consistent. We now always instantiate *all*
messages of a transaction and never fail during the operation. In case
"errorp" is given, it is set to the individual error code of each
instance, or 0 on success.

If CONTINUE is specified, any error possibly written into "errorp" is
ignored and does *NOT* abort the operation. If CONTINUE is *NOT*
specified, the last error that is non-zero is returned from the
operation and it is not committed.

If a management error happens *before* the transaction is started, or
after it was instantiated but before it was committed, it is returned
straight.

If, any any time, a fault occurs, EFAULT is returned instead of the
original error-code.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>