peer: merge peer_info into peer
The only reason to keep these separate was to be able to deallocate
the peer_info object at DISCONNECT. However, all real resources are
anyway released on DISCONNECT, so there is no real gain to this.

Merge the two objects and make the code significantly simpler.

Taken from David's cleanup patch.

Signed-off-by: Tom Gundersen <teg@jklm.no>