bus1: merge the bus2-branch into master
We started out with the goal to split peer-lock into local.lock and
data.lock. This took longer than expected, and we continued much of the
new features on the separate branch.

Merge everything back into master now, so master is the major
development branch again.

This is a bit disruptive, but should be the last major change.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>