pool: merge free and busy slices
Instead of having separate free and busy slices, merge them so that each slice
indicates both the busy memory and the trailing free memory.

The pool always contains a root slice, which has size 0 and only contains
padding. The pool holds the kernel reference to this slice, and it will
never be deleted.

All other slices must have size > 0, and these are all linked into the
'busy' tree, indexed by their offsets. Moreover, if a slices has trailing
padding, it is also linked into the 'padding' tree, indexed by the size
of the padding.

When allocating a new slice, we find the smallest possible padding that
will still fit the slice and insert the new slice at the beginning of
the padding.

When removing a slice, its size and padding is merged into the padding
of the previous slice in the pool.

Signed-off-by: Tom Gundersen <teg@jklm.no>