peer: ensure msg.destination refers to a peer local handle id
The struct bus1_cmd_recv.msg describing a message received on a peer
using the BUS1_CMD_RECV command includes a destination which is a handle
ID referring to the node a message was sent to.

In the current implementation the handle value given to the receiving
peer is the one used by the sending peer to initially send the message.

As handle values are peer local, this ID has no meaning for the receiving
peer and can not be used to find which node the message was sent to.

This commit fixes the remote handle ID logic in bus1_peer_new_message to
make sure that remote handles will be converted to peer local handle ID
values.

Signed-off-by: Timothée Ravier <tim@siosm.fr>