peer: fix id-namespace used to return destination IDs on recv
We pin the handle used to send a messages as m->dst. This handle is the
sender-handle, though. Therefore, m->dst->anchor is the receiving
handle. Make sure to use the anchor to retrieve the destination ID
during RECV. The caller of RECV could not make sense of sender IDs,
since they're in the wrong namespace.

Reported-by: Timothée Ravier <tim@siosm.fr>
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>