pool: don't allocate memory in the pool
The pool was managing the memory of each allocated slice. Instead, drop any
memory handling from the pool and let the caller pass in the required memory.

This means the pool no longer call kmalloc() under the data lock, and it allows
the memory used for each slice to be embedded in the corresponding message
object, removing the kmalloc() call alltogether.

This changes the lifetimes of both message objects and slices. Slices no
longer have user and kernel references. A slice can be marked as published
which means it cannot be deallocated, and it is up to the caller to ensure
this. When a slice is first published a reference is taken on the underlying
message object, and the reference is dropped after unpublishing the slice.
A slice is only deallocated after all the references to its message have
been dropped, guaranteeing that it is no longer published.

Before, message objects were released as soon as they were received by the
user, but now they are pinned until the user unpublishes the corresponding
slice.

Signed-off-by: Tom Gundersen <teg@jklm.no>