peer: drop support for PEEK
We have no compelling usecase for PEEK support at the moment. Drop
it for now to allow us to simplify the logic quite a bit.

Receiving the same message several times (which PEEK allowed) has
counterintuitive semantics in particular due to different fd numbers
being used each time and due to the accounting of handles.

Signed-off-by: Tom Gundersen <teg@jklm.no>