user: don't lock around charges/discharges
Each of the operations is atomic, though the charge as a whole no
longer is.

The effect of this is that it would now be possible for two
simultaneous charges to both fail, whereas (at most) one would have
succeeded before this change. This is an edge-case that we don't
particularly care to support.

Signed-off-by: Tom Gundersen <teg@jklm.no>