user: lock the limits object directly rather than the owning user
A limits object may either be associated with a user, or with a
peer, which in turn is associated with a user. In order to modify
the IDR in the limits object, the associated user was locked.
Change this to keep the lock in the limits object itself.

Not a functional change.

Signed-off-by: Tom Gundersen <teg@jklm.no>