user: fix a race in access to the usage objects
When charging a quota, we first fetched the current usage, then
determine if the quota can be charged, and then updated the usage.

This used to be protected by the user lock, but that broke in
6c4da221fab98441d757b5f59a82e9be90e3831d.

Reorder the logic so that we first update the usage object, then
use the result to determine if the quota can be charged. This means
that we could get some false negatives by charging the usage twice
at the same time, one of which will later be reverted as the quota
will fail. However, we do not attempt to protect against such
scenarios, and such a scenario can only be exploited by a user to
affect itself.

Signed-off-by: Tom Gundersen <teg@jklm.no>